pyproject.toml
[tool.poetry]
name = "analog-garage"
version = "0.1.0"
description = "SMS Simulation"
authors = ["Chirag Trivedi <chirag.trivedi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "0.19.2"
ruff = "0.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
